factor upload functionality out of app.py
